+++ projects/rocprofiler-compute/src/utils/utils.py
@@ -38,6 +38,7 @@ import tempfile
 import time
 from pathlib import Path as path
 from typing import Optional
+import uuid
 
 import pandas as pd
 import yaml
@@ -1640,3 +1641,6 @@ def parse_sets_yaml(arch):
         if set_option:
             sets_info[set_option] = set_item
     return sets_info
+
+def get_uuid(length=8):
+    return uuid.uuid4().hex[:length]